% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_envvars.r
\name{check_envvars}
\alias{check_envvars}
\title{Check that CONNECT_SERVER and CONNECT_API_KEY environment variables are set}
\usage{
check_envvars()
}
\description{
Check that CONNECT_SERVER and CONNECT_API_KEY environment variables are set
}
