% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pin.R
\name{create_pin}
\alias{create_pin}
\title{Create a pin on Connect}
\usage{
create_pin(
  x,
  name = deparse(substitute(x)),
  description = deparse(substitute(x)),
  server = Sys.getenv("CONNECT_SERVER"),
  key = Sys.getenv("CONNECT_API_KEY"),
  metadata = list(server_audit = TRUE),
  type = NULL
)
}
\arguments{
\item{x}{An R object (e.g., data frame, list, etc.).}

\item{server}{Connect server address.}

\item{key}{Administrator API key.}
}
\description{
Helper function used internally to pin audit data in a consistent manner.
}
