% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tag_pin.R
\name{create_tag_pin}
\alias{create_tag_pin}
\title{Create Tag for pinned content}
\usage{
create_tag_pin(pin_id, tag_name, conn = create_connection())
}
\arguments{
\item{pin_id}{guid for pin}

\item{tag_name}{tag schema}

\item{conn}{connection}
}
\value{

}
\description{
Create Tag for pinned content
}
