% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content_tbl.R
\name{get_content_tbl}
\alias{get_content_tbl}
\title{Get the Content table from the Connect server via \code{connectapi}}
\usage{
get_content_tbl(conn = create_connection())
}
\arguments{
\item{conn}{the Connect server connection details containing the server and API key}
}
\value{
a tibble
}
\description{
Get the Content table from the Connect server via \code{connectapi}
}
