% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_tbl.R
\name{get_group_names}
\alias{get_group_names}
\title{List the names of groups on the Connect server}
\usage{
get_group_names(x = groups_tbl)
}
\arguments{
\item{x}{groups_tbl}
}
\value{
a vector
}
\description{
List the names of groups on the Connect server
}
