% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_tbl.R
\name{get_groups_summary}
\alias{get_groups_summary}
\title{Get a Summary of the Groups Table}
\usage{
get_groups_summary(
  group_names_tbl = group_names_tbl,
  group_members_tbl = group_members_tbl
)
}
\arguments{
\item{groups_tbl}{The table that is generated by \code{get_groups_tbl()}}
}
\value{
a tibble
}
\description{
Get a Summary of the Groups Table
}
