% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_tbl.R
\name{get_groups_tbl}
\alias{get_groups_tbl}
\title{Get the Groups Table via Connect API}
\usage{
get_groups_tbl(conn = create_connection())
}
\arguments{
\item{conn}{the Connect server connection details containing the server and
API key}
}
\value{
a tibble
}
\description{
Get the Groups Table via Connect API
}
