% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lock_history_tbl.R
\name{get_lock_history_tbl}
\alias{get_lock_history_tbl}
\title{Get Locking event history from audit logs}
\usage{
get_lock_history_tbl(audit_logs)
}
\arguments{
\item{audit_logs}{audit logs}
}
\value{
a tibble
}
\description{
Get Locking event history from audit logs
}
