% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_current_tbl.R
\name{get_user_current_tbl}
\alias{get_user_current_tbl}
\title{Get current user table}
\usage{
get_user_current_tbl(user_historical_tbl = historical_users)
}
\arguments{
\item{user_historical_tbl}{the Historical Users Table}

\item{conn}{the Connect server connection details containing the server and API key}
}
\value{
a tibble
}
\description{
Get current user table
}
