% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_historical_tbl.R
\name{get_user_historical_tbl}
\alias{get_user_historical_tbl}
\title{Get historical user table based on Connect API user table}
\usage{
get_user_historical_tbl(users_tbl = users_tbl)
}
\arguments{
\item{users_tbl}{the Users Table via Connect API. By default the user table will be fetched via \code{get_users_tbl()}}
}
\value{
a tibble
}
\description{
Get historical user table based on Connect API user table
}
