% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_daysback.R
\name{get_users_daysback}
\alias{get_users_daysback}
\title{Get Users that have been active in the last \code{daysback} days}
\usage{
get_users_daysback(users_tbl, daysback)
}
\arguments{
\item{users_tbl}{Current Users table}

\item{daysback}{number of days back from today}
}
\value{
a tibble
}
\description{
Get Users that have been active in the last \code{daysback} days
}
