% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_role.R
\name{get_users_role}
\alias{get_users_role}
\title{Get Users by role}
\usage{
get_users_role(users_tbl, role)
}
\arguments{
\item{users_tbl}{Current Users table}

\item{role}{user role to filter}
}
\value{
a tibble
}
\description{
Get Users by role
}
