% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_tbl.R
\name{get_users_tbl}
\alias{get_users_tbl}
\title{Get the Users Table via the Connect API}
\usage{
get_users_tbl(conn = create_connection())
}
\arguments{
\item{conn}{the Connect server connection details containing the server and
API key.}
}
\value{
a tibble
}
\description{
This provides a tibble that relates to all users on Connect. It has the
following columns:
\itemize{
\item email:
\item username:
\item first_name:
\item last_name:
\item user_role:
\item created_time:
\item updated_time:
\item active_time:
\item confirmed:
\item locked:
\item guid:
}
}
