% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_tbl.R
\name{make_group_members_tbl}
\alias{make_group_members_tbl}
\title{Make a Summary Table of Group Members}
\usage{
make_group_members_tbl(x = "All")
}
\arguments{
\item{x}{Name of Group to list members of}
}
\value{
a reactable table
}
\description{
Make a Summary Table of Group Members
}
