% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_freshness.R
\name{pin_freshness_str}
\alias{pin_freshness_str}
\title{Pin Freshness String}
\usage{
pin_freshness_str(pin_freshness_tbl, pin_names = NULL)
}
\arguments{
\item{pin_freshness_tbl}{tibble of pin names and created date}

\item{pin_names}{if not NULL, subset of pin names from freshness tbl to extract into a string}
}
\value{
a string of pin names and timestamps
}
\description{
Pin Freshness String
}
