% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rscview_settings}
\alias{rscview_settings}
\title{Access to settings for the \strong{rscview} package}
\usage{
rscview_settings(admin_username, file = "rscview_settings.yml")
}
\arguments{
\item{admin_username}{The Connect username which has admin privileges.}

\item{file}{The filename for the YAML settings file.}
}
\description{
Access to settings for the \strong{rscview} package.
}
