% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_connection.R
\name{create_connection}
\alias{create_connection}
\title{Create connection to Connect server}
\usage{
create_connection(
  server = Sys.getenv("CONNECT_SERVER"),
  key = Sys.getenv("CONNECT_API_KEY")
)
}
\arguments{
\item{server}{Connect server address.}

\item{key}{Administrator API key.}
}
\value{
An RStudio Connect object that can be passed along to methods.
}
\description{
The \code{create_connection()} function is used to create a connection to your
Connect server via the \strong{connectapi} package.
}
