% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_current_tbl.R
\name{get_user_current_tbl}
\alias{get_user_current_tbl}
\title{Get current user table}
\usage{
get_user_current_tbl(
  conn = create_connection(),
  user_historical_tbl = get_user_historical_tbl()
)
}
\arguments{
\item{conn}{the Connect server connection details containing the server and API key}

\item{user_historical_tbl}{the Historical Users Table}
}
\value{
a tibble
}
\description{
Get current user table
}
