% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_historical_tbl.R
\name{get_users_tbl}
\alias{get_users_tbl}
\title{Get the Users Table via Connect API}
\usage{
get_users_tbl(conn = create_connection())
}
\arguments{
\item{conn}{the Connect server connection details containing the server and API key}
}
\value{
a tibble
}
\description{
Get the Users Table via Connect API
}
